+++ localization-finder-core/src/main/java/pl/tarko/software/finder/google/maps/service/impl/GoogleLocalizationServiceImpl.java
@@ -12,7 +12,7 @@ import org.springframework.stereotype.Service;
 import pl.tarko.software.finder.exceptions.GoogleLocalizationException;
 import pl.tarko.software.finder.google.maps.service.GoogleLocalizationService;
 import pl.tarko.software.finder.google.maps.util.LocalizationAdvancedGeoCoder;
-import pl.tarko.software.finder.lozalization.model.dto.LocalizationSearchParamDto;
+import pl.tarko.software.finder.localization.model.dto.LocalizationSearchParamDto;
 
 import java.io.IOException;
 import java.util.Optional;
