+++ localization-finder-core/src/main/java/pl/tarko/software/finder/localization/mapper/LocalizationMapperFactory.java
@@ -1,0 +1,9 @@
+package pl.tarko.software.finder.localization.mapper;
+
+import ma.glasnost.orika.impl.ConfigurableMapper;
+import org.springframework.stereotype.Component;
+
+@Component
+public class LocalizationMapperFactory extends ConfigurableMapper {
+
+}
