+++ localization-finder-core/src/main/java/pl/tarko/software/finder/localization/model/dto/BaseDistributionDto.java
@@ -1,0 +1,7 @@
+package pl.tarko.software.finder.localization.model.dto;
+
+import java.io.Serializable;
+
+public interface BaseDistributionDto extends Serializable {
+
+}
