+++ localization-finder-core/src/main/java/pl/tarko/software/finder/localization/model/dto/CountryDto.java
@@ -1,0 +1,12 @@
+package pl.tarko.software.finder.localization.model.dto;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Setter
+@Getter
+public class CountryDto implements BaseDistributionDto{
+
+    private Long id;
+    private String name;
+}
