+++ localization-finder-core/src/main/java/pl/tarko/software/finder/localization/model/jpa/LocatizationType.java
@@ -1,0 +1,5 @@
+package pl.tarko.software.finder.localization.model.jpa;
+
+public enum LocatizationType {
+    KINDER_GARTEN
+}
