+++ localization-finder-core/src/main/java/pl/tarko/software/finder/localization/repository/LocalizationRepository.java
@@ -1,0 +1,12 @@
+package pl.tarko.software.finder.localization.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import pl.tarko.software.finder.localization.model.jpa.Localization;
+
+public interface LocalizationRepository
+        extends JpaRepository<Localization, Long>/*
+                                                     * , QueryDslPredicateExecutor<
+													 * Localization>
+													 */ {
+
+}
