+++ localization-finder-core/src/main/java/pl/tarko/software/finder/localization/service/CountryService.java
@@ -1,0 +1,7 @@
+package pl.tarko.software.finder.localization.service;
+
+import pl.tarko.software.finder.localization.model.jpa.Country;
+
+public interface CountryService extends CrudService<Country, Long> {
+
+}
