+++ localization-finder-core/src/main/java/pl/tarko/software/finder/localization/service/TownService.java
@@ -1,0 +1,7 @@
+package pl.tarko.software.finder.localization.service;
+
+import pl.tarko.software.finder.localization.model.jpa.Town;
+
+public interface TownService extends CrudService<Town, Long> {
+
+}
