+++ localization-finder-core/src/main/java/pl/tarko/software/finder/localization/service/impl/CountryServiceImpl.java
@@ -1,0 +1,9 @@
+package pl.tarko.software.finder.localization.service.impl;
+
+import pl.tarko.software.finder.localization.model.dto.CountryDto;
+import pl.tarko.software.finder.localization.model.jpa.Country;
+import pl.tarko.software.finder.localization.service.CountryService;
+
+public class CountryServiceImpl extends CrudServiceImpl<CountryDto, Country, Long> implements CountryService {
+
+}
