+++ localization-finder-core/src/main/java/pl/tarko/software/finder/localization/service/impl/TownServiceImpl.java
@@ -1,0 +1,9 @@
+package pl.tarko.software.finder.localization.service.impl;
+
+import pl.tarko.software.finder.localization.model.dto.TownDto;
+import pl.tarko.software.finder.localization.model.jpa.Town;
+import pl.tarko.software.finder.localization.service.TownService;
+
+public class TownServiceImpl extends CrudServiceImpl<TownDto, Town, Long> implements TownService {
+
+}
