--- localization-finder-core/src/main/java/pl/tarko/software/finder/lozalization/mapper/LocalizationMapperFactory.java
@@ -1,9 +1,0 @@
-package pl.tarko.software.finder.lozalization.mapper;
-
-import ma.glasnost.orika.impl.ConfigurableMapper;
-import org.springframework.stereotype.Component;
-
-@Component
-public class LocalizationMapperFactory extends ConfigurableMapper {
-
-}
