--- localization-finder-core/src/main/java/pl/tarko/software/finder/lozalization/model/dto/BaseDistributionDto.java
@@ -1,7 +1,0 @@
-package pl.tarko.software.finder.lozalization.model.dto;
-
-import java.io.Serializable;
-
-public interface BaseDistributionDto extends Serializable {
-
-}
