--- localization-finder-core/src/main/java/pl/tarko/software/finder/lozalization/model/jpa/LocatizationType.java
@@ -1,5 +1,0 @@
-package pl.tarko.software.finder.lozalization.model.jpa;
-
-public enum LocatizationType {
-    KINDER_GARTEN
-}
