--- localization-finder-core/src/main/java/pl/tarko/software/finder/lozalization/repository/LocalizationRepository.java
@@ -1,12 +1,0 @@
-package pl.tarko.software.finder.lozalization.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import pl.tarko.software.finder.lozalization.model.jpa.Localization;
-
-public interface LocalizationRepository
-        extends JpaRepository<Localization, Long>/*
-                                                     * , QueryDslPredicateExecutor<
-													 * Localization>
-													 */ {
-
-}
