--- localization-finder-core/src/main/java/pl/tarko/software/finder/lozalization/service/CountryService.java
@@ -1,7 +1,0 @@
-package pl.tarko.software.finder.lozalization.service;
-
-import pl.tarko.software.finder.lozalization.model.jpa.Country;
-
-public interface CountryService extends CrudService<Country, Long> {
-
-}
