--- localization-finder-core/src/main/java/pl/tarko/software/finder/lozalization/service/CrudService.java
@@ -1,9 +1,0 @@
-package pl.tarko.software.finder.lozalization.service;
-
-import java.io.Serializable;
-
-public interface CrudService<T, ID extends Serializable> {
-    <S extends T> void save(S entity);
-
-    void delete(ID id);
-}
