--- localization-finder-core/src/main/java/pl/tarko/software/finder/lozalization/service/TownService.java
@@ -1,7 +1,0 @@
-package pl.tarko.software.finder.lozalization.service;
-
-import pl.tarko.software.finder.lozalization.model.jpa.Town;
-
-public interface TownService extends CrudService<Town, Long> {
-
-}
