--- localization-finder-core/src/main/java/pl/tarko/software/finder/lozalization/service/impl/CountryServiceImpl.java
@@ -1,9 +1,0 @@
-package pl.tarko.software.finder.lozalization.service.impl;
-
-import pl.tarko.software.finder.lozalization.model.dto.CountryDto;
-import pl.tarko.software.finder.lozalization.model.jpa.Country;
-import pl.tarko.software.finder.lozalization.service.CountryService;
-
-public class CountryServiceImpl extends CrudServiceImpl<CountryDto, Country, Long> implements CountryService {
-
-}
