--- localization-finder-core/src/main/java/pl/tarko/software/finder/lozalization/service/impl/TownServiceImpl.java
@@ -1,9 +1,0 @@
-package pl.tarko.software.finder.lozalization.service.impl;
-
-import pl.tarko.software.finder.lozalization.model.dto.TownDto;
-import pl.tarko.software.finder.lozalization.model.jpa.Town;
-import pl.tarko.software.finder.lozalization.service.TownService;
-
-public class TownServiceImpl extends CrudServiceImpl<TownDto, Town, Long> implements TownService {
-
-}
